# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FaceTrack API
  description: 'FaceTrack attendance system API (requirements §12). Portal: session
    cookie + CSRF header; integrations: X-API-Key.'
  version: 1.0.0
paths:
...
  /api/v1/events:
    get:
      tags:
      - recognition
      summary: Recognition events
      operationId: list_events
      parameters:
      - name: camera_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Camera Id
      - name: employee_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Employee Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/RecognitionStatus'
          - type: 'null'
          title: Status
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^-?[a-z_]+$
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_EventOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/events/{event_id}/void:
    post:
      tags:
      - recognition
      summary: Mark a false recognition
      description: 'FR-28: the event is voided (kept for threshold tuning) and the
        attendance day is rebuilt without it.'
      operationId: void_event
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventVoid'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EventOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/events/{event_id}/snapshot:
    get:
      tags:
      - recognition
      summary: Event face snapshot
      description: Allowed for event viewers, and for attendance viewers within their
        scope (check-in hover).
      operationId: event_snapshot
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/unknown-faces:
    get:
      tags:
      - recognition
      summary: Unknown-face review queue
      operationId: list_unknown_faces
      parameters:
      - name: review_status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ReviewStatus'
          - type: 'null'
          default: pending
          title: Review Status
      - name: camera_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Camera Id
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^-?[a-z_]+$
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_UnknownFaceOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/unknown-faces/{unknown_id}/snapshot:
    get:
      tags:
      - recognition
      summary: Unknown face snapshot
      operationId: unknown_snapshot
      parameters:
      - name: unknown_id
        in: path
        required: true
        schema:
          type: integer
          title: Unknown Id
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/unknown-faces/{unknown_id}/assign:
    post:
      tags:
      - recognition
      summary: Assign an unknown face to an employee
      description: 'FR-27: creates the attendance sighting; with `add_to_gallery`
        the snapshot is also enrolled after

        the FR-9 checks (requires consent).'
      operationId: assign_unknown
      parameters:
      - name: unknown_id
        in: path
        required: true
        schema:
          type: integer
          title: Unknown Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnknownFaceAssign'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UnknownFaceAssignResult_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/unknown-faces/{unknown_id}:
    patch:
      tags:
      - recognition
      summary: Dismiss
      operationId: update_unknown
      parameters:
      - name: unknown_id
        in: path
        required: true
        schema:
          type: integer
          title: Unknown Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnknownFaceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UnknownFaceOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_EventOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/EventOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[EventOut]
...
    ApiResponse_UnknownFaceAssignResult_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/UnknownFaceAssignResult'
      type: object
      required:
      - message
      - data
      title: ApiResponse[UnknownFaceAssignResult]
    ApiResponse_UnknownFaceOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/UnknownFaceOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[UnknownFaceOut]
...
    EventOut:
      properties:
        id:
          type: integer
          title: Id
        event_uuid:
          type: string
          title: Event Uuid
        camera_id:
          type: integer
          title: Camera Id
        camera_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Camera Name
        employee_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Employee Id
        employee_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Code
        employee_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Name
        status:
          $ref: '#/components/schemas/RecognitionStatus'
        confidence:
          type: number
          title: Confidence
        liveness_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Liveness Score
        track_id:
          type: integer
          title: Track Id
        captured_at:
          type: string
          title: Captured At
        snapshot_available:
          type: boolean
          title: Snapshot Available
        model_name:
          type: string
          title: Model Name
        void_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Void Reason
        voided_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Voided At
      type: object
      required:
      - id
      - event_uuid
      - camera_id
      - employee_id
      - status
      - confidence
      - liveness_score
      - track_id
      - captured_at
      - snapshot_available
      - model_name
      - void_reason
      - voided_at
      title: EventOut
    EventVoid:
      properties:
        reason:
          type: string
          maxLength: 500
          minLength: 3
          title: Reason
      additionalProperties: false
      type: object
      required:
      - reason
      title: EventVoid
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedResponse_EventOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/EventOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - message
      - data
      - pagination
      title: PaginatedResponse[EventOut]
...
    PaginatedResponse_UnknownFaceOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/UnknownFaceOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - message
      - data
      - pagination
      title: PaginatedResponse[UnknownFaceOut]
...
    Pagination:
      properties:
        page:
          type: integer
          minimum: 1.0
          title: Page
        page_size:
          type: integer
          minimum: 1.0
          title: Page Size
        total:
          type: integer
          minimum: 0.0
          title: Total
      type: object
      required:
      - page
      - page_size
      - total
      title: Pagination
...
    RecognitionStatus:
      type: string
      enum:
      - recognized
      - unknown
      - voided
      title: RecognitionStatus
      description: Status of a stored recognition event (§11). The engine only emits
        RECOGNIZED or UNKNOWN.
    ReviewStatus:
      type: string
      enum:
      - pending
      - assigned
      - dismissed
      title: ReviewStatus
      description: Review state of an unknown face (FR-27).
...
    UnknownFaceAssign:
      properties:
        employee_id:
          type: integer
          title: Employee Id
        add_to_gallery:
          type: boolean
          title: Add To Gallery
          default: false
      additionalProperties: false
      type: object
      required:
      - employee_id
      title: UnknownFaceAssign
    UnknownFaceAssignResult:
      properties:
        unknown_face:
          $ref: '#/components/schemas/UnknownFaceOut'
        attendance_updated:
          type: boolean
          title: Attendance Updated
        added_to_gallery:
          type: boolean
          title: Added To Gallery
        gallery_rejection_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Gallery Rejection Reason
      type: object
      required:
      - unknown_face
      - attendance_updated
      - added_to_gallery
      title: UnknownFaceAssignResult
    UnknownFaceOut:
      properties:
        id:
          type: integer
          title: Id
        recognition_event_id:
          type: integer
          title: Recognition Event Id
        captured_at:
          type: string
          title: Captured At
        camera_id:
          type: integer
          title: Camera Id
        camera_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Camera Name
        liveness_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Liveness Score
        review_status:
          $ref: '#/components/schemas/ReviewStatus'
        assigned_employee_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Employee Id
        reviewed_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Reviewed At
        snapshot_available:
          type: boolean
          title: Snapshot Available
        group_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Group Id
      type: object
      required:
      - id
      - recognition_event_id
      - captured_at
      - camera_id
      - liveness_score
      - review_status
      - assigned_employee_id
      - reviewed_at
      - snapshot_available
      title: UnknownFaceOut
    UnknownFaceUpdate:
      properties:
        review_status:
          type: string
          const: dismissed
          title: Review Status
      additionalProperties: false
      type: object
      required:
      - review_status
      title: UnknownFaceUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError